updated to have float data type and successfully made bullet travel diagonally

+++ SDL/Bullet.h
@@ -12,7 +12,7 @@ public:
 	Bullet(Vector2 position, Vector2 velocity, Vector2 size, SDL_Renderer* renderer, TextureManager* texture);
 	void reload();
 	void shoot(Vector2 playerPos);
-	void render(SDL_Renderer* renderer, const int camX, const int camY) const;
+	void render(SDL_Renderer* renderer, const float camX, const float camY) const;
 	void handleEvent(const SDL_Event& e, Vector2 playerPos);
 	void update();
 	void setTargetPos(const Vector2& targetPos);
@@ -29,7 +29,7 @@ private:
 	TextureManager* texture;
 
 	bool active;
-	int OFFSCREEN_BUFFER = 10;
+	float OFFSCREEN_BUFFER = 10;
 };
 
 #endif
